feat/fix: módulo MedicionCorporal completo + fix FetchType + fix logs
NUEVO MÓDULO - MedicionCorporal:
- MedicionCorporalMapper: fix target usuarioId (estaba mapeando a id)
- IMedicionCorporalService: interfaz con todos los métodos
- MedicionCorporalService: implementación completa con cálculo automático de IMC a partir de peso y altura
- MedicionCorporalController: endpoints REST completos
- IMedicionCorporalRepository: añadidos métodos findFirst, count,
  deleteBy y existsBy
- MedicionCorporalDTO: fix Double -> BigDecimal en peso y altura
- MedicionCorporalCreateDTO: fix Double -> BigDecimal en peso y altura

FIX ENTIDADES - FetchType.LAZY -> FetchType.EAGER:
- MedicionCorporal, Notificacion, ObjetivoPersonal, ProgresoEjercicio
- Comida, SesionEntrenamiento, Rutina, EjercicioRealizado
- RutinaEjercicio, AlimentoComida

FIX LOGS:
- logback-spring.xml: eliminado fichero activo gymprofit.log genérico
  que causaba rotaciones incorrectas con fechas anteriores
- Ahora cada día crea su propio fichero: gymprofit_YYYY-MM-DD.log

+++ api/gymprofit-api/src/main/java/com/gymprofit/api/dto/entity/medicioncorporal/MedicionCorporalDTO.java
@@ -15,8 +15,8 @@ public class MedicionCorporalDTO implements Serializable {
     private Integer id;
     private Integer usuarioId;
     private LocalDateTime fecha;
-    private Double peso;
-    private Double altura;
+    private BigDecimal peso;
+    private BigDecimal altura;
     private BigDecimal imc;
     private BigDecimal grasaCorporal;
     private BigDecimal masaMuscular;
